cli/requirements.txt
python-dotenv==0.19.2
click==8.0.3
colorama==0.4.4
termcolor==1.1.0